Cosmos: fix driver test compile under all-targets after #4588 API changes
The merge-preview against origin/main surfaced pre-existing breakage: #4588 renamed CosmosDriverRuntime::get_or_create_driver to create_driver (now taking only DriverOptions, account embedded), made PartitionFailoverOptions fields private, and removed the per-operation PPCB knobs from OperationOptions. Several inherited driver tests still used the old APIs and only compiled behind feature gates, so CI's --all-features --all-targets clippy fails.

Fixes: migrate the three create_driver call sites to the new single-arg signature; migrate the PPCB override tests to configure thresholds via driver-level PartitionFailoverOptions (new build_ppcb_fixture with a 1-failure threshold; build_fixture delegates with None so non-PPCB tests are unchanged); cast write_failure_threshold() (u32) to i32 for the write_failure_count field; add a test-only PartitionFailoverConfig alias to keep inherited operation_pipeline tests compiling.

+++ sdk/cosmos/azure_data_cosmos_driver/src/driver/pipeline/operation_pipeline.rs
@@ -5753,7 +5753,7 @@ mod tests {
         let pk_range_id: super::PartitionKeyRangeId = "0".parse().unwrap();
         let mut partitions = PartitionEndpointState::new(PartitionFailoverConfig::default());
         partitions.per_partition_circuit_breaker_enabled = true;
-        let write_threshold = partitions.config.write_failure_threshold;
+        let write_threshold = partitions.config.write_failure_threshold() as i32;
         partitions.circuit_breaker_overrides.insert(
             pk_range_id.clone(),
             PartitionFailoverEntry {
